- renamed bouncing dots property file - removed unnecessary control from the stroy
