+++ vue-components/src/components/bouncingDotsProps.ts
